[Static analysis] 2 duplicated blocks of code must be removed.
[Severity] Major
[Issue] #119 #206 #216 #221
